+++ src/index.html
@@ -7,21 +7,20 @@
     <title>To Do List</title>
 </head>
 <body>
-    <div class="main-Page">
-      <h1 class="tittle">To Do List</h1>
-      <div class="toDoList">
-        <div class="input">
-          <input class="task" type="text" placeholder="add to your list" />
-          <button ><i class=" add">add</i></button>
-        </div>
-        
-        <div class="listItem"></div>
-
-        <ul class="ul-items"></ul>
-        
-        
+  <div class=".main-Page">
+    <h1 class="header" Heading = "My To Do List">My To Do List</h1>
+    <div class="toDoList">
+      <div class="entryContainer">
+        <input class="toDoInput" type="text" placeholder="add your to do item here ..." />
+        <button class="addToDo"><i class="fa-solid fa-arrow-turn-down"></i></button>
       </div>
-      <button class="Completedbtn">Clear Completed Tasks</button>
+      
+      <div class="toDoListItemContainer"></div>
+      
+      
     </div>
+    <button class="clearCompleted">Clear Completed Tasks</button>
+  </div>
+  <script defer src="index.js"></script>
   </body>
 </html>
