+++ src/modules/clearAll.js
@@ -1,0 +1,19 @@
+import localGet from './storage.js';
+
+const clearCompleted = () => {
+  let inCompleted = localGet.filter((item) => {
+    if (!item.completed) {
+      return item;
+    }
+    return null;
+  });
+
+  inCompleted = inCompleted.map((item, id) => {
+    item.index = id + 1;
+    return item;
+  });
+  localStorage.setItem('listStorage', JSON.stringify(inCompleted));
+  window.location.reload();
+};
+
+export default clearCompleted;
