+++ src/modules/declarez.js
@@ -1,9 +1,9 @@
-const listChores = document.getElementById('todoList');
-const checkbox = document.querySelector('input[name=check]');
-const choreDesc = document.querySelector('.choreDesc');
-const addList = document.querySelector('input[name=addtext]');
-const clear = document.getElementById('clear');
+// const listChores = document.getElementById('todoList');
+// const checkbox = document.querySelector('input[name=check]');
+// const choreDesc = document.querySelector('.choreDesc');
+// const addList = document.querySelector('input[name=addtext]');
+// const clear = document.getElementById('clear');
 
-export {
-  listChores, checkbox, choreDesc, addList, clear,
-};
+// export {
+//   listChores, checkbox, choreDesc, addList, clear,
+// };
