+++ src/modules/storage.js
@@ -1,0 +1,3 @@
+const localGet = JSON.parse(localStorage.getItem('listStorage')) || [];
+
+export default localGet;
