+++ src/modules/templateList.js
@@ -1,0 +1,7 @@
+export default class ListTemplate {
+    constructor(description, completed, index) {
+      this.description = description;
+      this.completed = completed;
+      this.index = index;
+    }
+  }
